CheckSpeedchoiceOption returns an invalid value on invalid input

+++ include/speedchoice.h
@@ -40,7 +40,7 @@
  */
 struct OptionChoiceConfig
 {
-    s32 x; // do not store the Y coordinate. Y is automatically calculated depending on the row the option is placed and therefore is not necessary.
+    u8 x; // do not store the Y coordinate. Y is automatically calculated depending on the row the option is placed and therefore is not necessary.
     const u8 *string; // can be null
 };
